Variois fixes (mostly to placate gcc)
* examples/base64.c, examples/mta.c,
imap4d/auth_gss.c, imap4d/imap4d.c,
imap4d/preauth.c, libmu_auth/ldap.c,
libmu_auth/virtual.c, libmu_scm/mu_body.c,
libmu_scm/mu_port.c, libproto/mailer/smtp.c,
mailbox/acl.c, mailbox/secret.c, pop3d/pop3d.c,
python/libmu_py/address.c, sql/odbc.c: Fix argument
signedness.

* imap4d/fetch.c, imap4d/store.c (closures):
Change type of `count' to int, to match the
signature of util_msgset.
* include/mailutils/guile.h (mu_scm_message_get): Fix
return type (const is useless here).
* libmu_scm/mu_message.c (mu_scm_message_get): Likewise.
(scm_mu_message_copy): Fix type of the `wr' automatic variable.
* libmu_scm/mu_mime.c (mu_scm_mime_get): Remove const qualifier
from the return type.
* libmu_argp/cmdline.c: Include stdlib.h.

+++ libmu_scm/mu_body.c
@@ -108,7 +108,7 @@ SCM_DEFINE (scm_mu_body_read_line, "mu-body-read-line", 1, 0, 0,
 #define FUNC_NAME s_scm_mu_body_read_line
 {
   struct mu_body *mbp;
-  int n, nread;
+  size_t n, nread;
   int status;
   
   SCM_ASSERT (mu_scm_is_body (BODY), BODY, SCM_ARG1, FUNC_NAME);
